Shift Wizard components into separate files
