+++ src/components/Wizard/WizardStep.tsx
@@ -1,0 +1,11 @@
+export interface IWizardStepProps {
+  id: string
+  label: string
+  optional?: boolean
+}
+
+export const WizardStep: React.FunctionComponent<IWizardStepProps> = ({
+  children,
+}) => {
+  return <>{children}</>
+}
